+++ rbx_types/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased Changes
 * `Ref` can now represent null explicitly via `Ref::none` and `Ref::is_none`.
+* Added `Region3` and `Region3int16` to `Variant` and `VariantType`.
 
 ## 0.1.0 (2020-02-05)
 * Initial release of types crate, should have rough feature parity with rbx_dom_weak.